# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/discount-cards": {
            "get": {
                "tags": [
                    "Discount Cards"
                ],
                "summary": "Get all discount cards",
                "operationId": "ffa3fec293bc17b0d674d1f43613aa68",
                "responses": {
                    "200": {
                        "description": "List of discount cards"
                    }
                }
            },
            "post": {
                "tags": [
                    "Discount Cards"
                ],
                "summary": "Create a new discount card",
                "operationId": "361ebe2b8c6861ed3bc0148efa8f31bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "zlava_id",
                                    "zlavovy_kod"
                                ],
                                "properties": {
                                    "zlava_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "zlavovy_kod": {
                                        "type": "string",
                                        "example": "DISCOUNT2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Discount card created"
                    }
                }
            }
        },
        "/api/discount-cards/{id}": {
            "get": {
                "tags": [
                    "Discount Cards"
                ],
                "summary": "Get a specific discount card",
                "operationId": "abc317b96739547b9287297b2255fc46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount card details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Discount Cards"
                ],
                "summary": "Update a discount card",
                "operationId": "c16da8855477c25f77d50adadb1cb565",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "zlava_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "zlavovy_kod": {
                                        "type": "string",
                                        "example": "DISCOUNT2025"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discount card updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Discount Cards"
                ],
                "summary": "Delete a discount card",
                "operationId": "82942751a54951a03ab12658c5bc8060",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount card",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount card deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Zľavová karta bola vymazaná."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/discounts": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Get all discounts",
                "operationId": "eaa03209dcdee72030b74953ddb84bee",
                "responses": {
                    "200": {
                        "description": "List of discounts"
                    }
                }
            },
            "post": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Create a new discount",
                "operationId": "9fad542c7104853a4bcb49d059f4d242",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nazov",
                                    "vyska"
                                ],
                                "properties": {
                                    "nazov": {
                                        "type": "string",
                                        "example": "Summer Sale"
                                    },
                                    "vyska": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 20.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Discount created"
                    }
                }
            }
        },
        "/api/discounts/{id}": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Get a specific discount",
                "operationId": "00178dc9f7f203e911afc5909d70ee68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Update a discount",
                "operationId": "7d4f99b630ac17ab9ed0a5f7d0ccf888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nazov": {
                                        "type": "string",
                                        "example": "Winter Sale"
                                    },
                                    "vyska": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Discount updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Delete a discount",
                "operationId": "c5d8f8e842a2dca9bd212b0945b44762",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the discount",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Zľava bola vymazaná"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}